Prevent reinstallation of an already in-use group key
Track the current GTK and IGTK that is in use and when receiving a
(possibly retransmitted) Group Message 1 or WNM-Sleep Mode Response, do
not install the given key if it is already in use. This prevents an
attacker from trying to trick the client into resetting or lowering the
sequence counter associated to the group key.

Change-Id: I5a315accb15ca5d2464acc3c370c66e07d57c7ea
Merged-In: I5a315accb15ca5d2464acc3c370c66e07d57c7ea
Signed-off-by: Mathy Vanhoef <Mathy.Vanhoef@cs.kuleuven.be>
Signed-off-by: Glen Kuhne <kuh@google.com>
Bug: 65245581
Test: WiFi integration tests
Merged-In: I1273cf9ee58154b16512ef4284a8d2c5541d3c01